fixed getssrProps
