+++ src/pages/api/top.ts
@@ -1,4 +1,3 @@
-// Next.js API route support: https://nextjs.org/docs/api-routes/introduction
 import type { NextApiRequest, NextApiResponse } from 'next'
 
 type Data = {
@@ -9,5 +8,7 @@ export default function handler(
   req: NextApiRequest,
   res: NextApiResponse<Data>,
 ) {
+  if (req.method !== 'GET') res.status(405).end()
+
   res.status(200).json({ name: 'John Doe' })
 }
